Update README.md
The software aims to assist people living in shared accomodations to keep an overview of
their flatmates, their finances and shared responsibilities. Each user of the program can
create their own account, enter their individual expenses and enter a list of purchased
groceries. This information is shared with the other users and allows them to keep track of
whether one member of the household is spending disproportionate amounts of money or
if the financial situation is balanced. A shared grocery list eases individual shopping and
helps to plan meals.